# frontend/packages/api-client/openapi.yaml
openapi: 3.1.0
info:
  title: HireLens API
  version: 0.0.1
paths:
...
  /api/identity/users:
    get:
      operationId: listUsers
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    User:
      type: object
      required: [id, tenantId, externalSubject, displayName, roles, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        externalSubject:
          type: string
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    CreateUser:
      type: object
      required: [externalSubject, displayName, roles]
      properties:
        externalSubject:
          type: string
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string